# api-proxy/openapi.yaml
openapi: 3.0.0
info:
  title: Vestix API Proxy
  description: API Proxy for the Vestix-App, aggregating various financial data sources.
  version: 1.0.0

paths:
  /markets:
    get:
      summary: Get market data for a specific category
      parameters:
        - name: category
          in: query
          required: true
          schema:
            type: string
            enum: [index, metal, fx, crypto, future, status]
      responses:
        '200':
          description: A list of tickers for the specified category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UnifiedTicker'

  /ticker:
    get:
      summary: Get detailed ticker information for a specific symbol
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detailed ticker information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnifiedTicker'

  /kline:
    get:
      summary: Get Kline (candlestick) data for a symbol
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: interval
          in: query
          required: true
          schema:
            type: string
            enum: ["1m", "5m", "15m", "1h", "4h", "1d", "1w", "1M"]
      responses:
        '200':
          description: A list of Kline data points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UnifiedKline'

  /search:
    get:
      summary: Search for assets
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UnifiedTicker'

components:
  schemas:
    UnifiedTicker:
      type: object
      properties:
        symbol:
          type: string
          example: 'BTC'
        name:
          type: string
          example: 'Bitcoin'
        category:
          type: string
          enum: [index, metal, fx, crypto, future]
        price:
          type: number
          example: 68000.50
        change:
          type: number
          example: 1500.00
        changePercent:
          type: number
          example: 2.25
        timestamp:
          type: number
          example: 1678886400000
    UnifiedKline:
      type: object
      properties:
        time:
          type: number
          example: 1678886400000
        open:
          type: number
          example: 67000.00
        high:
          type: number
          example: 68500.00
        low:
          type: number
          example: 66800.00
        close:
          type: number
          example: 68000.50
        volume:
          type: number
          example: 1000000.00